Изменение:
- Добавлена pb и домейн модель Snapshot
-- Теперь вместо отдельных файлов передается снапшот
-- Для FileEntry добавлено поле SubtreeSize
-- В FileManager теперь собирается количество файлов в поддеревьях и возвращать снапшот по файлам
-- Заменены все сигнатуры методов на актуальные

+++ internal/domain/file_metadata.go
@@ -1,7 +1,7 @@
 package domain
 
 type FileMetadata struct {
-	ModifiedUnix uint32
-	SizeBytes    uint32
+	ModifiedUnix uint64
+	SizeBytes    uint64
 	IsDirectory  bool
 }
